Update associated media in core.md
Extended the media object definition and the associated media section. Consider any item within a record as a media object, but express it with one of the three main media objects (DataDownload, ImageObject, or VideoObject) rather than just using mediaObject. The encodingFormat and contentUrl are properties of a media object. This means we cannot directly have them if there is no AssociatedMedia > MediaObject in the metadata. So, I did not add them to the table. Instead, I defined them in the AssociatedMedia section.